handful new gnucash transactions
Originally I added these transactions before the conversion to sqlite.
But since I wanted to commit the sqlite conversion as a standalone
change, this required a script to reconsile the two sqlite dumps before
and after adding these transactions.

The script handles a bunch of things:

(1) unstable INSERT statement order in the sqlite dump
(2) unstable slot and currency guids in the sqlite data
(3) unstable slot sequence ids in the sqlite data

Features (2) and (3) will probably not be useful in the future, because
ids and guids should be stable as long as there I don't do any
conversions to/from sqlite format in the future. So I'll remove this
handling from the script in a followup commit.

Feature (1) might probably be useful in the future, because INSERT
statement order could easy change in future sqlite dumps.